lab4 Designer: use doubles instead of integers for drawing
And stream instead of arguments vector

+++ lab4/Designer/Common/App/ICanvas.h
@@ -7,7 +7,7 @@ class ICanvas
 public:
 	virtual void SetColor(Color color) = 0;
 	virtual void DrawLine(Point const& from, Point const& to) = 0;
-	virtual void DrawEllipse(Point const& center, int horizontalRadius, int verticalRadius) = 0; // TODO: this + coords make double
+	virtual void DrawEllipse(Point const& center, double horizontalRadius, double verticalRadius) = 0;
 
 	virtual ~ICanvas() = default;
 };
