+++ lab4/Designer/Common/Infrastructure/SVG/CCanvas.h
@@ -11,7 +11,7 @@ public:
 
 	void SetColor(Color color) override;
 	void DrawLine(const Point& from, const Point& to) override;
-	void DrawEllipse(const Point& center, int horizontalRadius, int verticalRadius) override;
+	void DrawEllipse(const Point& center, double horizontalRadius, double verticalRadius) override;
 
 private:
 	std::ostream& m_output;
